requirements.txt
Django>=5.2.4,<5.3
djangorestframework>=3.16.0,<3.17.0
psycopg2>=2.9.10,<2.10
drf-spectacular>=0.28.0,<0.29.0